# docsclear/swagger.yaml
basePath: /
definitions:
...
  entity.Currency:
    enum:
    - XOF
    - XAF
    type: string
    x-enum-comments:
      CurrencyXAF: Franc CFA (CEMAC)
      CurrencyXOF: Franc CFA (UEMOA)
    x-enum-descriptions:
    - Franc CFA (UEMOA)
    - Franc CFA (CEMAC)
    x-enum-varnames:
    - CurrencyXOF
    - CurrencyXAF
...
  entity.PaymentProvider:
    enum:
    - orange_money
    - moov_money
    - wave
    - yenga_pay
    - cash
    - mock
    type: string
    x-enum-varnames:
    - ProviderOrangeMoney
    - ProviderMoovMoney
    - ProviderWave
    - ProviderYengaPay
    - ProviderCash
    - ProviderMock
  entity.PaymentStatus:
    enum:
    - pending
    - processing
    - success
    - failed
    - refunded
    - cancelled
    - expired
    type: string
    x-enum-comments:
      PaymentStatusCancelled: Annulé
      PaymentStatusExpired: Expiré
      PaymentStatusFailed: Échoué
      PaymentStatusPending: Créé, en attente
      PaymentStatusProcessing: En cours de traitement
      PaymentStatusRefunded: Remboursé
      PaymentStatusSuccess: Réussi
    x-enum-descriptions:
    - Créé, en attente
    - En cours de traitement
    - Réussi
    - Échoué
    - Remboursé
    - Annulé
    - Expiré
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusProcessing
    - PaymentStatusSuccess
    - PaymentStatusFailed
    - PaymentStatusRefunded
    - PaymentStatusCancelled
    - PaymentStatusExpired
...
  paymentdto.CompletePaymentRequest:
    properties:
      otp:
        type: string
      payment_id:
        type: string
    type: object
  paymentdto.InitiatePaymentRequest:
    properties:
      callback_url:
        type: string
      customer_email:
        type: string
      description:
        type: string
      metadata:
        additionalProperties: true
        type: object
      order_id:
        type: string
      phone_number:
        type: string
      provider:
        $ref: '#/definitions/entity.PaymentProvider'
    type: object
  paymentdto.InitiatePaymentResponse:
    properties:
      expires_at:
        type: integer
      message:
        type: string
      metadata:
        additionalProperties: true
        type: object
      payment_id:
        type: string
      provider_ref:
        type: string
      redirect_url:
        type: string
      status:
        $ref: '#/definitions/entity.PaymentStatus'
      ussd_code:
        type: string
    type: object
  paymentdto.PaymentResponse:
    properties:
      amount_cents:
        type: integer
      completed_at:
        type: string
      created_at:
        type: string
      currency:
        $ref: '#/definitions/entity.Currency'
      customer_phone:
        type: string
      description:
        type: string
      id:
        type: string
      initiated_at:
        type: string
      order_id:
        type: string
      provider:
        $ref: '#/definitions/entity.PaymentProvider'
      provider_ref:
        type: string
      status:
        $ref: '#/definitions/entity.PaymentStatus'
    type: object
  paymentdto.RefundPaymentRequest:
    properties:
      amount_cents:
        type: integer
      payment_id:
        type: string
      reason:
        type: string
    type: object
  paymentdto.WebhookResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  utils.AppError:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@goshop.dev
    name: API Support
  description: GoShop - E-commerce API with observability and metrics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: 1.0.0
paths:
...
  /api/orders/{id}/pay:
    post:
      consumes:
      - application/json
      description: Lance le processus de paiement pour une commande donnée via le
        fournisseur de paiement choisi.
      parameters:
      - description: ID de la commande (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Détails du paiement (provider, phone_number, etc.)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/paymentdto.InitiatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/paymentdto.InitiatePaymentResponse'
        "400":
          description: Payload invalide ou commande non éligible
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Commande introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "409":
          description: Un paiement est déjà en cours pour cette commande
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Initier un paiement pour une commande
      tags:
      - Payments
...
  /api/payments:
    get:
      consumes:
      - application/json
      description: Retourne la liste paginée et filtrée des paiements pour la boutique
        active.
      parameters:
      - description: 'Filtrer par statut de paiement (ex: success, pending, failed)'
        in: query
        name: status
        type: string
      - description: 'Filtrer par fournisseur (ex: wave, orange_money)'
        in: query
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/paymentdto.PaymentResponse'
            type: array
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Lister les paiements
      tags:
      - Payments
  /api/payments/{id}:
    get:
      consumes:
      - application/json
      description: Retourne le statut et les informations complètes d'une transaction
        de paiement spécifique.
      parameters:
      - description: ID du paiement (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentdto.PaymentResponse'
        "400":
          description: ID de paiement manquant
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Paiement introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Récupérer les détails d'un paiement
      tags:
      - Payments
  /api/payments/{id}/complete:
    post:
      consumes:
      - application/json
      description: 'Finalise un paiement qui nécessite une validation OTP (ex: certains
        flux Mobile Money).'
      parameters:
      - description: ID du paiement (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Code OTP de validation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/paymentdto.CompletePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentdto.PaymentResponse'
        "400":
          description: Payload invalide, OTP incorrect ou opération non supportée
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "403":
          description: Le paiement n'appartient pas à votre boutique
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Paiement introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Compléter un paiement en deux étapes (OTP)
      tags:
      - Payments
  /api/payments/{id}/refund:
    post:
      consumes:
      - application/json
      description: Lance une demande de remboursement pour une transaction de paiement
        réussie.
      parameters:
      - description: ID du paiement (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Détails du remboursement (montant, motif)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/paymentdto.RefundPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentdto.PaymentResponse'
        "400":
          description: Payload invalide ou remboursement impossible
          schema:
            $ref: '#/definitions/utils.AppError'
        "401":
          description: Non autorisé
          schema:
            $ref: '#/definitions/utils.AppError'
        "403":
          description: Le paiement n'appartient pas à votre boutique
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Paiement introuvable
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Rembourser un paiement
      tags:
      - Payments
...
  /webhooks/{provider}:
    post:
      consumes:
      - application/json
      description: Endpoint public sécurisé pour recevoir les notifications de changement
        de statut de paiement des fournisseurs (Wave, Orange, Moov, YengaPay).
      parameters:
      - description: Code du fournisseur (wave, orange_money, moov_money, yenga_pay)
        in: path
        name: provider
        required: true
        type: string
      - description: Signature HMAC-SHA256 du payload
        in: header
        name: X-Signature
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Webhook reçu et traité avec succès
          schema:
            $ref: '#/definitions/paymentdto.WebhookResponse'
        "400":
          description: Fournisseur inconnu ou signature invalide
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Recevoir les webhooks de paiement
      tags:
      - Webhooks
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"